pyproject.toml
[project]
name = "bad-autoencoding"
version = "0.1.0"
description = "Optical Context Compression Is Just (Bad) Autoencoding - Code for reproducing experiments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "addict>=2.4.0",
    "bitsandbytes>=0.48.2",
    "datasets>=4.3.0",
    "easydict>=1.13",
    "einops>=0.8.1",
    "huggingface-hub>=0.36.0",
    "jieba>=0.42.1",
    "matplotlib>=3.10.7",
    "nltk>=3.9.2",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "polars>=1.34.0",
    "pyarrow>=22.0.0",
    "rich>=14.2.0",
    "seaborn>=0.13.2",
    "tokenizers==0.20.3",
    "torch==2.6.0",
    "tqdm>=4.66.0",
    "torchvision>=0.21.0",
    "transformers==4.46.3",
    "wandb>=0.22.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
